requirements.txt
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
convertapi==2.0.0
cryptography==46.0.4
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
emoji==2.15.0
etelemetry==0.3.1
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
langdetect==1.0.9
llvmlite==0.46.0
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
olefile==0.47
openai==2.16.0
packaging==26.0
pandas==3.0.0
pathlib==1.0.1
postgrest==2.27.2
propcache==0.4.1
prov==2.1.1
psutil==7.2.2
puremagic==1.30
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.6.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pyxnat==1.6.4
RapidFuzz==3.14.3
rdflib==7.5.0
realtime==2.27.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
scipy==1.17.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tools==1.0.23
tqdm==4.67.2
traits==7.1.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
unstructured==0.18.31
unstructured-client==0.42.8
urllib3==2.6.3
uvicorn==0.40.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.1.1
yarl==1.22.0
